// src/schemas/schemas.ts
// Generated by ts-to-zod
import { z } from "zod";

export const ClientSchema = z.object({
  age: z.coerce
    .number({
      invalid_type_error: "Невалидна стойност",
    })
    .optional(),
  name: z
    .string({
      required_error: "Задължително поле",
      invalid_type_error: "Невалидна стойност",
    })
    .trim()
    .min(1, "Задължително поле"),
  fee: z.coerce
    .number({
      invalid_type_error: "Невалидна стойност",
    })
    .optional(),
  occupation: z
    .string({
      invalid_type_error: "Невалидна стойност",
    })
    .trim()
    .optional(),
  facts: z
    .string({
      required_error: "Задължително поле",
      invalid_type_error: "Невалидна стойност",
    })
    .trim()
    .optional(),
  phone: z
    .string({
      invalid_type_error: "Невалидна стойност",
    })
    .trim()
    .optional(),
  email: z
    .string({
      invalid_type_error: "Невалидна стойност",
    })
    .trim()
    .optional()
});